SecuritySettingsScreen refactor
